requirements.txt
tensorflow
flask
numpy
pillow
werkzeug
h5py
